// GiveAway-Node/src/module/user/dto/enum.ts
export enum UserRole {
    ADMIN = 'Admin',
    CUSTOMER = 'Customer',
    OWNER = 'Owner',
    COMPANY = "Company",
    USER = "IndividualUser"
}

export enum AuthType {
    LOCAL = 'Local',
    FACEBOOK = 'Facebook',
    GOOGLE = 'Google',
    APPLE = 'Apple'
}


export enum AppType {
    APP = 'App',
    BROWSER = 'Browser'
}

export enum OsType {
    ANDROID = 'Android',
    IOS = 'Ios',
    WINDOWS = 'Windows'
}

export enum PlaylistType {
    OTHER = 'Other',
    CURATED = 'Curated',
    AUDIENCE = 'Audience'
}

export enum ArtistType {
    OTHER = 'Other',
    COMPANY = 'Company'
}

export enum SearchType {
    SONG = 'SONG',
    ALBUM = 'ALBUM',
    ARTIST = 'ARTIST'
}

export enum TransectionType {
    DEBIT = 'Debit',
    CREDIT = 'Credit'
}